+++ android/library/src/main/java/com/liferay/mobile/screens/thingscreenlet/model/WorkflowTask.kt
@@ -20,27 +20,54 @@ import com.liferay.apio.consumer.model.Relation
 import com.liferay.apio.consumer.model.Thing
 import com.liferay.apio.consumer.model.get
 import com.liferay.mobile.screens.R
-import com.liferay.mobile.screens.thingscreenlet.screens.views.Detail
 import com.liferay.mobile.screens.thingscreenlet.screens.views.Row
 import com.liferay.mobile.screens.thingscreenlet.screens.views.Scenario
 import java.util.*
 
-data class BlogPosting(
-	val headline: String?,
-	val alternativeHeadline: String?,
-	val articleBody: String?,
-	val creator: Relation?,
-	val createDate: Date?,
-	val type: String?) {
+data class WorkflowTask(
+	val completed: Boolean,
+	val dateCreated: Date?,
+	val name: String,
+	val expires: Date?,
+	val comment: Comment?,
+	val blogPost: BlogPosting?) {
 
 	companion object {
 		val DEFAULT_VIEWS: MutableMap<Scenario, Int> =
 			mutableMapOf(
-				Detail to R.layout.blog_posting_detail_default,
-				Row to R.layout.blog_posting_row_default
+				Row to R.layout.workflow_task_row_default
 			)
 
 		val converter: (Thing) -> Any = {
+
+			val completed = it["completed"] as Boolean
+
+			val dateCreated = (it["dateCreated"] as? String)?.asDate()
+
+			val name = it["name"] as String
+
+			val expires = (it["expires"] as? String)?.asDate()
+
+			val comment = (it["comment"] as? Relation)?.let {
+				getComment(it)
+			}
+
+			val blogPost = (it["blogPost"] as? Relation)?.let {
+				getBlogPost(it)
+			}
+
+			WorkflowTask(completed, dateCreated, name, expires, comment, blogPost)
+		}
+
+		private fun getComment(it: Relation): Comment? {
+			val text = it["text"] as? String
+
+			val type = graph[it.id]?.value?.type?.get(0)
+
+			return Comment(text, type)
+		}
+
+		private fun getBlogPost(it: Relation): BlogPosting? {
 
 			val headline = it["headline"] as? String
 
@@ -54,7 +81,7 @@ data class BlogPosting(
 
 			val type = graph[it.id]?.value?.type?.get(0)
 
-			BlogPosting(headline, alternativeHeadline, articleBody, creator, dateCreated, type)
+			return BlogPosting(headline, alternativeHeadline, articleBody, creator, dateCreated, type)
 		}
 	}
 }