+++ android/library/src/main/java/com/liferay/mobile/screens/thingscreenlet/screens/adapter/ThingAdapter.kt
@@ -19,13 +19,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.github.kittinunf.result.success
 import com.liferay.apio.consumer.ApioConsumer
+import com.liferay.apio.consumer.delegates.convert
+import com.liferay.apio.consumer.fetch
+import com.liferay.apio.consumer.model.Thing
+import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.extensions.inflate
 import com.liferay.mobile.screens.thingscreenlet.model.Collection
-import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.screens.views.BaseView
 import com.liferay.mobile.screens.thingscreenlet.screens.views.Scenario
 import com.liferay.apio.consumer.delegates.convert
-import com.liferay.apio.consumer.model.Thing
 import okhttp3.HttpUrl
 
 class ThingAdapter(collection: Collection, val listener: Listener) :
@@ -38,7 +40,6 @@ class ThingAdapter(collection: Collection, val listener: Listener) :
 
 	val totalItems = collection.totalItems
 	val members = collection.members?.toMutableList() ?: mutableListOf()
-
 	val nextPage = collection.pages?.next
 
 	private val apioConsumer = ApioConsumer()