+++ android/library/src/main/java/com/liferay/mobile/screens/thingscreenlet/screens/views/ThingView.kt
@@ -19,11 +19,11 @@ import android.util.AttributeSet
 import android.view.View
 import android.widget.LinearLayout
 import android.widget.TextView
+import com.liferay.apio.consumer.delegates.observeNonNull
+import com.liferay.apio.consumer.model.Thing
 import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.delegates.bind
 import com.liferay.mobile.screens.thingscreenlet.screens.ThingScreenlet
-import com.liferay.apio.consumer.delegates.observeNonNull
-import com.liferay.apio.consumer.model.Thing
 
 class ThingView @JvmOverloads constructor(
 	context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0, defStyleRes: Int = 0) : BaseView,