+++ android/library/src/main/java/com/liferay/mobile/screens/thingscreenlet/screens/views/detail/BlogPostingDetailView.kt
@@ -18,14 +18,14 @@ import android.content.Context
 import android.util.AttributeSet
 import android.widget.RelativeLayout
 import android.widget.TextView
+import com.liferay.apio.consumer.delegates.converter
+import com.liferay.apio.consumer.model.Thing
+import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.delegates.bindNonNull
 import com.liferay.mobile.screens.thingscreenlet.model.BlogPosting
-import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.screens.ThingScreenlet
 import com.liferay.mobile.screens.thingscreenlet.screens.events.Event
 import com.liferay.mobile.screens.thingscreenlet.screens.views.BaseView
-import com.liferay.apio.consumer.delegates.converter
-import com.liferay.apio.consumer.model.Thing
 import com.liferay.mobile.screens.util.AndroidUtil
 
 class BlogPostingDetailView @JvmOverloads constructor(