+++ android/library/src/main/java/com/liferay/mobile/screens/thingscreenlet/screens/views/detail/CollectionDetailView.kt
@@ -15,13 +15,14 @@
 package com.liferay.mobile.screens.thingscreenlet.screens.views.detail
 
 import android.content.Context
+import android.support.v7.widget.DividerItemDecoration
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
 import android.view.View
-import android.widget.Button
 import android.widget.FrameLayout
-import android.widget.RelativeLayout
+import com.liferay.apio.consumer.delegates.converter
+import com.liferay.apio.consumer.model.Thing
 import com.liferay.mobile.screens.R
 import com.liferay.mobile.screens.thingscreenlet.delegates.bindNonNull
 import com.liferay.mobile.screens.thingscreenlet.model.Collection
@@ -30,8 +31,6 @@ import com.liferay.mobile.screens.thingscreenlet.screens.adapter.ThingAdapter
 import com.liferay.mobile.screens.thingscreenlet.screens.events.Event
 import com.liferay.mobile.screens.thingscreenlet.screens.views.BaseView
 import com.liferay.mobile.screens.thingscreenlet.screens.views.Scenario
-import com.liferay.apio.consumer.delegates.converter
-import com.liferay.apio.consumer.model.Thing
 
 open class CollectionDetailView @JvmOverloads constructor(
 	context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0, defStyleRes: Int = 0) : BaseView,
@@ -42,6 +41,7 @@ open class CollectionDetailView @JvmOverloads constructor(
 	val recyclerView by bindNonNull<RecyclerView>(R.id.collection_recycler_view)
 
 	override var thing: Thing? by converter<Collection> {
+		recyclerView.addItemDecoration(DividerItemDecoration(getContext(), DividerItemDecoration.VERTICAL));
 		recyclerView.layoutManager = LinearLayoutManager(context)
 		recyclerView.adapter = ThingAdapter(it, this)
 	}